Change the data structure of the Instance, Net and Cluster objects.
Instance:
	- There is now a dictionary in the Graph object with all the instances of the design
	- Each instance contains references to the Cluster and Net objects it belongs to.
Net: When parsing the net-instance file, we fetch the Cluster reference from the instance object and add it to the Net object.
This way, the net is already aware of all the clusters it's connecting. Doing so allows us to drasticaly reduce the loops and time necessary to build the hyperedges.

In practice, we divide the time by a factor 10.

The next step is to improve the merging part of the hypergraph building, as it still takes around 75% of the total execution time (the other 25% being the metis execution with the tested design: LDPC with 100 custom clusters).