Definitive choice on hyperedges merging algorithm.
After timing some possibilities, it turns out that the double loop with breaks is the fastest way to go.
symmetric_difference is extremely slow and not reliable enough.

In the end, we do not gain so much time beyond the 'x10' acceleration factor we had some commits ago.
Still, it's taking only a few seconds when it was taking minutes before.