Handles the memory blocks again.
This had been broken by the change in data structures in the classes.
In particular, keep in mind that when you create a cluster, instances should be created at the same time, and when you create an Instance, the Cluster it belongs to which should be referenced as well.
This is crucial for the buildHyperedge function. Cross-references are life, cross-references are love.