New script: 3D manhattan.
This script extracts information from files generated by the def_parser and the phoney partitioner.
From there, it computes the 2D manhattan distance between each pair of gates separated.
Then, based on its average, Q1 and Q3, tells wht should the 3D width be in order to not increase that distance when stacking the two dies.

It should be easy to use from another script, as the 'main' function only parses the cli arguments and forward them to the 'Evaluate3DLength(...)' function.