Phoney: MAJOR UPDATE: Pins.
Read a pin file to create extra clusters assigned to the bottom die with no area.
Their purpose is to take into account that extra connection, i.e. cell to pin, which was overlooked previously but should not have been.

This update is critical and should render most previous partitioning results void.